Try code with 10K resistor
https://hackaday.io/project/8435-2-analog-inputs-for-esp8266-without-multiplexer
Firstly I hooked up the joystick as shown in the schematic.
The two 10k resistors coming from Joystick X and Y are there so when the
digital GPIO is sinked to GND the Analog input is not also sinked completely
to ground. to check for X GPIO 13 is set to floating and GPIO 14 is set to GND.
using those 2 resistors along the voltage at the ADC will be half
of what is coming from the Joystick X which will be a maximum of 1.65V.
Since the ADC will only read to a maximum of 1 Volt, this
voltage needs to be brought down even more.
An additional 10K resistor is used to bring the voltage down even further.
Ideally this should be a smaller value but in testing 10K seemed to be adequate
and this further brought down the BOM.
Using this code I was able to make a function to calibrate the
joystick by finding it's maximum and minimum values for both
the X and Y and then creating functions for getting the X and the Y values.

Signed-off-by: Andrey Shamis <andrey.shamis@intel.com>